requirements.txt
numpy~=1.17.2
pillow~=6.2.0